+++ webapi/src/main/scala/org/knora/webapi/responders/v2/SearchResponderV2.scala
@@ -53,6 +53,7 @@ import org.knora.webapi.messages.v2.responder.resourcemessages.*
 import org.knora.webapi.slice.admin.domain.model.KnoraProject.ProjectIri
 import org.knora.webapi.slice.admin.domain.model.User
 import org.knora.webapi.slice.ontology.domain.service.IriConverter
+import org.knora.webapi.slice.ontology.domain.service.OntologyRepo
 import org.knora.webapi.slice.ontology.repo.service.OntologyCache
 import org.knora.webapi.store.triplestore.api.TriplestoreService
 import org.knora.webapi.store.triplestore.api.TriplestoreService.Queries.Construct
@@ -96,12 +97,18 @@ trait SearchResponderV2 {
     query: ConstructQuery,
     schemaAndOptions: SchemaRendering,
     user: User,
+    limitToProject: Option[ProjectIri] = None,
   ): Task[ReadResourcesSequenceV2]
-
-  def gravsearchV2(query: IRI, rendering: SchemaRendering, user: User): Task[ReadResourcesSequenceV2] = for {
-    q <- ZIO.attempt(GravsearchParser.parseQuery(query))
-    r <- gravsearchV2(q, rendering, user)
-  } yield r
+  def gravsearchV2(
+    query: IRI,
+    rendering: SchemaRendering,
+    user: User,
+    limitToProject: Option[ProjectIri],
+  ): Task[ReadResourcesSequenceV2] =
+    for {
+      q <- ZIO.attempt(GravsearchParser.parseQuery(query))
+      r <- gravsearchV2(q, rendering, user, limitToProject)
+    } yield r
 
   /**
    * Performs a count query for a Gravsearch query provided by the user.
@@ -110,12 +117,28 @@ trait SearchResponderV2 {
    * @param user  the client making the request.
    * @return a [[ResourceCountV2]] representing the number of resources that have been found.
    */
-  def gravsearchCountV2(query: ConstructQuery, user: User): Task[ResourceCountV2]
-  def gravsearchCountV2(query: IRI, user: User): Task[ResourceCountV2] =
+  def gravsearchCountV2(query: ConstructQuery, user: User, limitToProject: Option[ProjectIri]): Task[ResourceCountV2]
+  def gravsearchCountV2(query: IRI, user: User, limitToProject: Option[ProjectIri]): Task[ResourceCountV2] =
     for {
       q <- ZIO.attempt(GravsearchParser.parseQuery(query))
-      r <- gravsearchCountV2(q, user)
+      r <- gravsearchCountV2(q, user, limitToProject)
     } yield r
+
+  /**
+   * Performs a Gravsearchquery to find resources that link to the specified resource.
+   *
+   * @param resourceIri the IRI of the resource to which incoming links are to be found.
+   * @param offset      the offset to be used for paging.
+   * @param rendering   the schema of the response.
+   * @param user        the client making the request.
+   */
+  def searchIncomingLinksV2(
+    resourceIri: IRI,
+    offset: Int,
+    rendering: SchemaRendering,
+    user: User,
+    limitToProject: Option[ProjectIri],
+  ): Task[ReadResourcesSequenceV2]
 
   /**
    * Performs a fulltext search and returns the resources count (how many resources match the search criteria),
@@ -228,10 +251,63 @@ final case class SearchResponderV2Live(
   private val stringFormatter: StringFormatter,
   private val iriConverter: IriConverter,
   private val constructTransformer: ConstructTransformer,
+  private val ontologyRepo: OntologyRepo,
 ) extends SearchResponderV2
     with LazyLogging {
 
   private implicit val sf: StringFormatter = stringFormatter
+
+  /**
+   * Performs a Gravsearchquery to find resources that link to the specified resource.
+   *
+   * @param resourceIri the IRI of the resource to which incoming links are to be found.
+   * @param offset      the offset to be used for paging.
+   * @param rendering   the schema of the response.
+   * @param user        the client making the request.
+   */
+  def searchIncomingLinksV2(
+    resourceIri: IRI,
+    offset: Int,
+    rendering: SchemaRendering,
+    user: User,
+    limitToProject: Option[ProjectIri],
+  ): Task[ReadResourcesSequenceV2] = {
+    val query: String =
+      s"""
+         |PREFIX knora-api: <http://api.knora.org/ontology/knora-api/simple/v2#>
+         |
+         |CONSTRUCT {
+         |?incomingRes knora-api:isMainResource true .
+         |
+         |?incomingRes ?incomingProp <$resourceIri> .
+         |
+         |} WHERE {
+         |
+         |?incomingRes a knora-api:Resource .
+         |
+         |?incomingRes ?incomingProp <$resourceIri> .
+         |
+         |<$resourceIri> a knora-api:Resource .
+         |
+         |?incomingProp knora-api:objectType knora-api:Resource .
+         |
+         |knora-api:isRegionOf knora-api:objectType knora-api:Resource .
+         |knora-api:isPartOf knora-api:objectType knora-api:Resource .
+         |
+         |FILTER NOT EXISTS {
+         |?incomingRes  knora-api:isRegionOf <$resourceIri> .
+         |}
+         |
+         |FILTER NOT EXISTS {
+         |?incomingRes  knora-api:isPartOf <$resourceIri> .
+         |?incomingRes knora-api:seqnum ?seqnum .
+         |}
+         |
+         |} OFFSET $offset
+         |""".stripMargin
+
+    gravsearchV2(query, rendering, user, limitToProject)
+  }
 
   /**
    * Performs a fulltext search and returns the resources count (how many resources match the search criteria),
@@ -415,7 +491,11 @@ final case class SearchResponderV2Live(
     } yield apiResponse
   }
 
-  override def gravsearchCountV2(query: ConstructQuery, user: User): Task[ResourceCountV2] =
+  override def gravsearchCountV2(
+    query: ConstructQuery,
+    user: User,
+    limitToProject: Option[ProjectIri],
+  ): Task[ResourceCountV2] =
     for {
       _ <- // make sure that OFFSET is 0
         ZIO
@@ -457,7 +537,9 @@ final case class SearchResponderV2Live(
         )
 
       ontologiesForInferenceMaybe <-
-        inferenceOptimizationService.getOntologiesRelevantForInference(query.whereClause)
+        limitToProject.fold(
+          inferenceOptimizationService.getOntologiesRelevantForInference(query.whereClause),
+        )(getProjectOntologies)
 
       countQuery <- queryTraverser.transformSelectToSelect(
                       inputQuery = prequery,
@@ -482,6 +564,7 @@ final case class SearchResponderV2Live(
     query: ConstructQuery,
     schemaAndOptions: SchemaRendering,
     user: User,
+    limitToProject: Option[ProjectIri] = None,
   ): Task[ReadResourcesSequenceV2] = {
 
     for {
@@ -495,20 +578,24 @@ final case class SearchResponderV2Live(
       // Create a Select prequery
       querySchema <-
         ZIO.fromOption(query.querySchema).orElseFail(AssertionException(s"InputQuery has no querySchema"))
-      gravsearchToPrequeryTransformer <- ZIO.attempt(
-                                           new GravsearchToPrequeryTransformer(
-                                             constructClause = query.constructClause,
-                                             typeInspectionResult = typeInspectionResult,
-                                             querySchema = querySchema,
-                                             appConfig = appConfig,
-                                           ),
-                                         )
+
+      gravsearchToPrequeryTransformer <-
+        ZIO.attempt(
+          new GravsearchToPrequeryTransformer(
+            constructClause = query.constructClause,
+            typeInspectionResult = typeInspectionResult,
+            querySchema = querySchema,
+            appConfig = appConfig,
+          ),
+        )
 
       // TODO: if the ORDER BY criterion is a property whose occurrence is not 1, then the logic does not work correctly
       // TODO: the ORDER BY criterion has to be included in a GROUP BY statement, returning more than one row if property occurs more than once
 
       ontologiesForInferenceMaybe <-
-        inferenceOptimizationService.getOntologiesRelevantForInference(query.whereClause)
+        limitToProject.fold(
+          inferenceOptimizationService.getOntologiesRelevantForInference(query.whereClause),
+        )(getProjectOntologies)
 
       prequery <-
         queryTraverser.transformConstructToSelect(
@@ -656,17 +743,18 @@ final case class SearchResponderV2Live(
           ZIO.succeed(Map.empty[IRI, MappingAndXSLTransformation])
         }
 
-      apiResponse <- constructResponseUtilV2.createApiResponse(
-                       mainResourcesAndValueRdfData = mainQueryResults,
-                       orderByResourceIri = mainResourceIris,
-                       pageSizeBeforeFiltering = pageSizeBeforeFiltering,
-                       mappings = mappingsAsMap,
-                       queryStandoff = queryStandoff,
-                       versionDate = None,
-                       calculateMayHaveMoreResults = true,
-                       targetSchema = schemaAndOptions.schema,
-                       requestingUser = user,
-                     )
+      apiResponse <-
+        constructResponseUtilV2.createApiResponse(
+          mainResourcesAndValueRdfData = mainQueryResults,
+          orderByResourceIri = mainResourceIris,
+          pageSizeBeforeFiltering = pageSizeBeforeFiltering,
+          mappings = mappingsAsMap,
+          queryStandoff = queryStandoff,
+          calculateMayHaveMoreResults = true,
+          versionDate = None,
+          targetSchema = schemaAndOptions.schema,
+          requestingUser = user,
+        )
     } yield apiResponse
   }
 
@@ -1045,6 +1133,18 @@ final case class SearchResponderV2Live(
       results = SparqlSelectResultBody(prequeryRowsMerged),
     )
   }
+
+  /**
+   * Returns the set of ontologies of a given project to which the search is limited.
+   *
+   * @param projectIri the IRI of the project.
+   * @return the set of ontology IRIs of the project.
+   */
+  private def getProjectOntologies(projectIri: ProjectIri): Task[Option[Set[SmartIri]]] =
+    ontologyRepo.findByProject(projectIri).map { ontologies =>
+      val ontologyIris = ontologies.map(_.ontologyMetadata.ontologyIri)
+      Option.when(ontologyIris.nonEmpty)(ontologyIris.toSet)
+    }
 }
 
 object SearchResponderV2Live {